requirements.txt

opencv-python-headless
numpy>=1.24.0
onnxruntime>=1.16.0
streamlit
scipy
pillow
